moved special attacks into separate module files
