+++ lib/dps/Dps.js
@@ -3,9 +3,10 @@ import { MaxHit } from "./MaxHit.js";
 import { Accuracy } from "./Accuracy.js";
 import { AttackSpeed } from "./AttackSpeed.js";
 import { SpecialWeapons } from "./SpecialWeapons.js";
-import { SpecialAttacks } from "./SpecialAttacks.js";
 import { HitFreqStore } from "./HitFreqStore.js";
 import Player from "../Player.js";
+
+import applySpecialAttack from "./SpecialAttacks.js";
 
 export class Dps {
 	constructor (stateObj) {
@@ -265,10 +266,9 @@ export class Dps {
 			"Enchanted pearl bolts fiery"
 		];
 		const specs = new SpecialWeapons(this.state, this.calcs);
-		const specialAttacks = new SpecialAttacks(this.state, this.calcs);
 
-		if (this.calcs.specName !== null) {
-			this.calcs = specialAttacks.applySpec(this.calcs.specName);
+		if (this.calcs.specName) {
+			this.calcs = applySpecialAttack(this.calcs.specName, this.calcs, this.state);
 		}
 
 		if (this.calcs.flags.includes("Scythe of vitur")) {
