+++ lib/dps/SpecialWeapons.js
@@ -623,7 +623,6 @@ export class SpecialWeapons {
 		hitStore.store([0], 1 - acc);
 
 		for (let h1 = min; h1 <= newMax; h1 += 1) {
-			// kerisHitList.push({ dmg: [h1], p: acc * 50 / 51 / (newMax + 1) })
 			hitStore.store([h1], acc / (newMax - min + 1));
 		}
